// src/entities/enums.ts
export enum UserRoles {
  ADMIN = 'ADMIN',
  AUTHOR = 'AUTHOR',
  STUDENT = 'STUDENT'
}

export enum StepCategories {
  LECTURE = 'LECTURE',
  READING = 'READING',
  QUIZ = 'QUIZ'
}

export enum QuestionCategories {
  SINGLE_CHOICE = 'SINGLE_CHOICE',
  MULTI_CHOICE = 'MULTI_CHOICE'
}

export enum TimePeriods {
  DAY = 'DAY',
  WEEK = 'WEEK',
  MONTH = 'MONTH'
}